Vim/Tmux: Add headless clipboard support
Using X11 forwarding with SSH to sync the system clipboard is overkill
and can be quite slow. The headless clipboard uses the tmux buffer to
share contents. Tmux 3.2 can even send the contents with an xterm escape
sequence to the terminal emulator and therefore update the clipboard on
the SSH client side. The only downside is, that the remote side cannot
read the contents from the SSH client side, so you have to explicitly
paste using the terminal emulator key binding, e.g. CTRL-Shift-V.