Vim: Save y<C-G> to system/tmux clipboard
The mapping is from fugitive.vim, but only uses setreg to store the git
object. Most of the time I want to use the path or commit outside of the
current Vim instance (e.g. browser or git command).

Fixed by calling the existing clipboard integration with the fugitive
object string to save it to the tmux buffer and send it to the terminal
emulator.